requirements.txt
torch
torchvision
numpy < 2.0.0
PyMCubes
matplotlib
imageio
tabulate